examples: fix example invocation in README
The example was moved into its own crate, but the README was not updated to
reflect this change.

Remove the link name attribute while we are at it. It is not needed, as the cc
crate will automatically make the compiled symbols available to the rest of the
Rust code.